requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Environment & Config
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Utilities
httpx==0.27.0
email-validator==2.1.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
